requirements.txt
openai
tiktoken
pypdf
pinecone-client
langchain
streamlit
supabase